feat: refactor type layout primitives
* Set up a separate `layout` module in the `miden-hir-type` crate which
  holds the `TypeRepr` implementation, as well as all of its supporting
  functionality.
* Provide an `Alignable` trait which can be used to calculate alignment
  offsets, next aligned value, and next multiple of any unsigned
  primitive integral type
* Rework the layout functions to use `Alignable`
* Rework code that duplicates functionality provided by `Alignable` to
  use the trait instead
